+++ Repositories/Implementations/UserRepository.cs
@@ -10,14 +10,17 @@ namespace Job_Portal.Repositories.Implementations
     public class UserRepository : IUserRepository
     {
         private readonly AppDbContext _context;
+
         public UserRepository(AppDbContext context)
         {
             _context = context;
         }
+
         public async Task<IEnumerable<User>> GetAllAsync()
         {
             return await _context.Users.ToListAsync();
         }
+
         public async Task<User> GetByIdAsync(int id)
         {
             var user = await _context.Users.FindAsync(id);
@@ -27,20 +30,29 @@ namespace Job_Portal.Repositories.Implementations
             }
             return user;
         }
-        public async Task AddAsync(User user)
+
+        public async Task<User> AddAsync(User user)
         {
             await _context.Users.AddAsync(user);
             await _context.SaveChangesAsync();
+            return user;
         }
-        public void Update(User user)
+
+        public async Task<User> UpdateAsync(User user)
         {
             _context.Users.Update(user);
-            _context.SaveChangesAsync();
+            await _context.SaveChangesAsync();
+            return user;
         }
-        public void Delete(User user)
+
+        public async Task DeleteAsync(int id)
         {
-            _context.Users.Remove(user);
-            _context.SaveChangesAsync();
+            var user = await _context.Users.FindAsync(id);
+            if (user != null)
+            {
+                _context.Users.Remove(user);
+                await _context.SaveChangesAsync();
+            }
         }
     }
 }
